+++ server/src/test/component_tests/api/v1/create-ride.test.ts
@@ -1,15 +1,29 @@
 import app from '../../../../app';
-import createUserAndLogin from './create-user-and-login.test';
 import { Ride, User, Verification } from '../../../../models';
 
+const uuid = require('uuid');
 const request = require('supertest');
+
+const baseUrl = '/api/v1';
 
 jest.mock('../../../../lib/nexmo', () => ({
   sendSms: jest.fn(() => true),
 }));
 
+const createUserAndLogin = async () => {
+  const operationId = uuid.v4();
+  await User.create({
+    phoneNumber: '972501234567', firstName: 'GUY', lastName: 'Serfaty', operationId,
+  });
+  await Verification.create({ phoneNumber: '972501234567', externalCode: '1234', operationId });
+  const response = await request(app).post(`${baseUrl}/login/vert`)
+    .send({ phoneNumber: '972501234567', code: '1234' })
+    .set('x-loco-op-id', operationId);
+  console.log('response', response.body);
+  return response.body;
+};
+
 describe('Create rides', () => {
-  const baseUrl = '/api/v1';
 
   beforeEach(async () => {
     await User.destroy({ truncate: true, force: true });
